Develop

+++ App/Services/JobService.cs
@@ -12,7 +12,7 @@ namespace App.Service
         private readonly IUserRepository _userRepository;
         private readonly IResponseBuilder _responseBuilder;
 
-        public JobService (IJobRepository jobRepository, IUserRepository userRepository, IResponseBuilder responseBuilder)
+        public JobService(IJobRepository jobRepository, IUserRepository userRepository, IResponseBuilder responseBuilder)
         {
             _jobRepository = jobRepository;
             _userRepository = userRepository;
@@ -30,10 +30,20 @@ namespace App.Service
                 var modality = await _jobRepository.GetModalityById(JobDTO.modality);
                 var status = await _jobRepository.GetStatusById(JobDTO.status);
 
-                Console.WriteLine(modality);
-                Console.WriteLine(status);
-                Console.WriteLine(user);
-                
+                if (user == null)
+                {
+                    return _responseBuilder.Conflict($"Nenhuma referencia do objeto usuario encontrada");
+                }
+
+                if (modality == null)
+                {
+                    return _responseBuilder.Conflict($"Nenhuma referencia do objeto modality encontrada");
+                }
+
+                if (status == null)
+                {
+                    return _responseBuilder.Conflict($"Nenhuma referencia do objeto status encontrada");
+                }
 
                 JobVacancy job = new()
                 {
@@ -51,19 +61,24 @@ namespace App.Service
             }
             catch (Exception ex)
             {
-                return _responseBuilder.InternalError( $"Ocorreu um erro interno: {ex.Message}");
+                return _responseBuilder.InternalError($"Ocorreu um erro interno: {ex.Message}");
             }
         }
 
-        public async Task<ResponseDTO> Delete(SearchByIdDTO searchByIdDTO)
+        public async Task<ResponseDTO> Delete(UserActionDTO userAction)
         {
             try
             {
-                var job = await _jobRepository.GetById(searchByIdDTO.Id);
+                var job = await _jobRepository.GetById(userAction.JobId);
 
-                if(job == null)
+                if (job == null)
                 {
                     return _responseBuilder.Conflict("Vaga não encontrada!");
+                }
+
+                if (job.UserId != userAction.UserId)
+                {
+                    return _responseBuilder.Conflict("Usuário não possui permissão!");
                 }
 
                 await _jobRepository.Delete(job);
@@ -72,26 +87,119 @@ namespace App.Service
             }
             catch (Exception ex)
             {
-                return _responseBuilder.InternalError( $"Ocorreu um erro interno: {ex.Message}");
+                return _responseBuilder.InternalError($"Ocorreu um erro interno: {ex.Message}");
             }
         }
 
-        public async Task<ResponseDTO> GetAllById(SearchByIdDTO searchByIdDTO)
+        public async Task<ResponseDTO> UpdateJob(UpdateJobDTO updateJobDTO)
         {
             try
             {
-                var jobs = await _jobRepository.GetAllById(searchByIdDTO.Id);
+                var job = await _jobRepository.GetById(updateJobDTO.jobId);
 
-                if(jobs == null)
+                if (job == null)
+                {
+                    return _responseBuilder.Conflict("Vaga não encontrada!");
+                }
+
+                if (job.UserId != updateJobDTO.userId)
+                {
+                    return _responseBuilder.Conflict("Usuário não possui permissão!");
+                }
+
+                job.Title = updateJobDTO.title;
+                job.Link = updateJobDTO.link;
+                job.EnterpriseName = updateJobDTO.enterpriseName;
+                job.ModalityId = updateJobDTO.modality;
+                job.VacancyStatusId = updateJobDTO.status;
+                job.UpdatedAt = DateTime.Now;
+
+                await _jobRepository.Edit(job);
+
+                return _responseBuilder.OK(job, "Vaga editada com sucesso!");
+
+            }
+            catch (Exception ex)
+            {
+                return _responseBuilder.InternalError($"Ocorreu um erro interno: {ex.Message}");
+            }
+        }
+
+        public async Task<ResponseDTO> GetAllById(Guid searchById)
+        {
+            try
+            {
+                List<JobResponseDTO> jobslist = new List<JobResponseDTO>();
+
+                var jobs = await _jobRepository.GetAllById(searchById);
+
+                if (jobs == null)
                 {
                     return _responseBuilder.NotFound("Nenhuma vaga foi encontrada!");
                 }
 
-                return _responseBuilder.OK(jobs, "Todos as vagas foram encontradas");
+                foreach (var item in jobs)
+                {
+                    var status = await _jobRepository.GetStatusById(item.VacancyStatusId);
+                    var modality = await _jobRepository.GetModalityById(item.ModalityId);
+
+                    jobslist.Add(new JobResponseDTO
+                    {
+                        id = item.Id,
+                        title = item.Title,
+                        link = item.Link,
+                        enterpriseName = item.EnterpriseName,
+                        status = status.Name,
+                        modality = modality.Name,
+                        createdAt = item.CreatedAt,
+                        updatedAt = item.UpdatedAt
+                    });
+                }
+
+                return _responseBuilder.OK(jobslist, "Todos as vagas foram encontradas");
             }
             catch (Exception ex)
             {
-                return _responseBuilder.InternalError( $"Ocorreu um erro interno: {ex.Message}");
+                return _responseBuilder.InternalError($"Ocorreu um erro interno: {ex.Message}");
+            }
+        }
+
+        public async Task<ResponseDTO> GetJobById(UserActionDTO jobByIdDto)
+        {
+            try
+            {
+                var job = await _jobRepository.GetById(jobByIdDto.JobId);
+
+                if (job == null)
+                {
+                    return _responseBuilder.NotFound("Nenhuma vaga foi encontrada!");
+                }
+
+                if (job.UserId != jobByIdDto.UserId)
+                {
+                    return _responseBuilder.Conflict("Usuário não possui permissão!");
+                }
+
+                var status = await _jobRepository.GetStatusById(job.VacancyStatusId);
+                var modality = await _jobRepository.GetModalityById(job.ModalityId);
+
+                JobResponseByIdDTO jobInfo = new()
+                {
+                    id = job.Id,
+                    title = job.Title,
+                    link = job.Link,
+                    enterpriseName = job.EnterpriseName,
+                    status = status.Id,
+                    modality = modality.Id,
+                    createdAt = job.CreatedAt,
+                    updatedAt = job.UpdatedAt
+                };
+
+                return _responseBuilder.OK(jobInfo, "Vaga encontrada com sucesso!");
+            }
+            catch (Exception ex)
+            {
+                return _responseBuilder.InternalError($"Ocorreu um erro interno: {ex.Message}");
             }
         }
 
@@ -99,7 +207,7 @@ namespace App.Service
         {
             try
             {
-                var statusList = _jobRepository.AllStatus();
+                var statusList = await _jobRepository.AllStatus();
 
                 VacancyStatus status = new()
                 {
@@ -112,7 +220,7 @@ namespace App.Service
             }
             catch (Exception ex)
             {
-                return _responseBuilder.InternalError( $"Ocorreu um erro interno: {ex.Message}");
+                return _responseBuilder.InternalError($"Ocorreu um erro interno: {ex.Message}");
             }
         }
 
@@ -120,6 +228,8 @@ namespace App.Service
         {
             try
             {
+                List<AllStatusResponseDTO> allStatus = new List<AllStatusResponseDTO>();
+
                 var status = await _jobRepository.AllStatus();
 
                 if (status == null)
@@ -127,11 +237,20 @@ namespace App.Service
                     return _responseBuilder.NotFound("Nenhum dado foi encontrado!");
                 }
 
-                return _responseBuilder.OK(status, "Dados encontrados com sucesso!");
+                foreach (var item in status)
+                {
+                    allStatus.Add(new AllStatusResponseDTO
+                    {
+                        id = item.Id,
+                        name = item.Name
+                    });
+                }
+
+                return _responseBuilder.OK(allStatus, "Dados encontrados com sucesso!");
             }
             catch (Exception ex)
             {
-                return _responseBuilder.InternalError( $"Ocorreu um erro interno: {ex.Message}");
+                return _responseBuilder.InternalError($"Ocorreu um erro interno: {ex.Message}");
             }
         }
 
@@ -152,7 +271,7 @@ namespace App.Service
             }
             catch (Exception ex)
             {
-                return _responseBuilder.InternalError( $"Ocorreu um erro interno: {ex.Message}");
+                return _responseBuilder.InternalError($"Ocorreu um erro interno: {ex.Message}");
             }
         }
 
@@ -160,6 +279,8 @@ namespace App.Service
         {
             try
             {
+                List<AllModalityResponseDTO> allModality = new List<AllModalityResponseDTO>();
+
                 var modalityList = await _jobRepository.AllModality();
 
                 if (modalityList == null)
@@ -167,11 +288,20 @@ namespace App.Service
                     return _responseBuilder.NotFound("Nenhum dado foi encontrado!");
                 }
 
-                return _responseBuilder.OK(modalityList, "Dados encontrados com sucesso!");
+                foreach (var item in modalityList)
+                {
+                    allModality.Add(new AllModalityResponseDTO
+                    {
+                        id = item.Id,
+                        name = item.Name
+                    });
+                }
+
+                return _responseBuilder.OK(allModality, "Dados encontrados com sucesso!");
             }
             catch (Exception ex)
             {
-                return _responseBuilder.InternalError( $"Ocorreu um erro interno: {ex.Message}");
+                return _responseBuilder.InternalError($"Ocorreu um erro interno: {ex.Message}");
             }
         }
     }
